Phase 21 follow-up: extract staging, asset registration, drop legacy log
- runtime/staging.py: RemoteStager owns the staging cache, in-flight
  ref deduplication, stage_task_inputs, stage_remote_value, and
  cache_lookup_requires_staging. Module-level helpers cover
  count_remote_inputs, remote_value_requires_staging,
  resolve_staging_jobs, and load_remote_publisher.

- runtime/artifacts/asset_registration.py: AssetRegistrar materialises
  AssetResult sentinels into AssetRefs, registering each version
  against the asset store and propagating lineage from upstream
  refs. Also exposes asset_index_for/render_asset_ref helpers.

- runtime/artifacts/output_index.py: artifact_index() (the typed
  per-task output summary used by provenance and lifecycle events)
  now lives next to the other artifact code.

- evaluator.py: drops from 2283 to 1782 lines (75 -> 61 methods).
  Constructs the new helpers in __post_init__ and delegates from
  every former call site. The dead _emit_legacy_log path is gone -
  the CLI long ago moved to RichEventRenderer subscribed to the
  event bus, and the duplicate stderr JSON path was only kept alive
  by tests that grepped capsys.err.

- evaluate(): now accepts an event_bus= kwarg so ad-hoc and test
  callers can observe lifecycle events without spinning up a CLI.

- tests/conftest.py: new EventCollector helper that subscribes a
  list collector to a fresh EventBus and exposes started/staging/
  cached/succeeded/failed accessors.

- tests: legacy stderr-grep assertions in test_evaluator,
  test_cache_integrity, test_remote_evaluator,
  test_warm_run_optimizations, and test_vw_pixi rewritten to use
  the EventCollector. Two _staging_cache test patches are now on
  evaluator._stager._staging_cache.
